BUG#37026691 Performance schema unit tests fail on Windows
This is a bug in the unit test code only,
production code is not affected.

Problem:
========

Performance schema unit tests fails on windows.
The error reported is that memory allocated with _aligned_malloc()
is freed with free() instead of _aligned_free().

Root cause:
===========

The fix for:
  Bug 36921583 PERFORMANCE_SCHEMA UNIT TESTS FAIL WITH UBSAN
was incomplete.

In particular, the unit test stub for pfs_free() was overlooked.

Fix:
====

Use the proper free() function in the unit test stubs to free aligned memory.

Change-Id: I5ba62d2ecfff19ef6d004f07dbe6a65bd7a44014