Bug#37008930:Assertion `strategy ==              Subquery_strategy::CANDIDATE_FOR_IN2EXISTS_OR_MAT' failed.
For a subquery which has window functions in select list, IN2EXISTS
transformation cannot happen. So subquery_strategy() for such a
subquery is always "MATERIALIZATION". However, this is true only when
the subquery is IN/ALL/ANY/SOME. While checking if materialization is
forced for a subquery, if the query has a IN subquery with its where
clause having an EXISTS subquery with window functions in it,
IsForcedMaterialization() was returning true. This is because the
innermost subquery having window functions always
returned its strategy as "MATERIALIZATION". However, for the case in
bugpage, strategy for IN subquery was already picked as IN2EXISTS. As a
result some checks failed when this subquery was considered for
materialization.
Fix is to add a check for quantified comparison predicate in
IsForcedMaterialization() so that the subquery with window function
in this case is not looked into. This is in sync with the check present in
FinalizeMaterialzedSubqueries().

Change-Id: I6f44488723de2485a5cb886411e46a01448d65f1