Bug#36902116 crash after walk_join_list()
A transform of a scalar subquery was attemped after a merge of derived
table containing the scalar subquery in its where clause. This is
unexpected and caused a crash.

Normally, if the subquery were a candidate for transform, it would
have happened before the merge of the derived table, but in this case
it wasn't transformed because it had an outer aggregation.  The
base_query block was inner to the merged query block, and
aggr_query_block was outer to the query block that was merged into.

The fact of the  outer aggregation is originally recorded in

  Item_singlerow_subselect::m_used_tables_cache, and in
  Item_singlerow_subselect::m_subquery_used_tables

via the bit OUTER_REF_TABLE_BIT. However, when performing
fix_after_pullout, this flag wasn't updated even though the aggregate
still has an outer aggregation, so the subquery was no longer barred
from an attemped transform.

Solution: make Item_sum::Item_sum::fix_after_pullout refresh this bit
for Item_sum::fix_after_pullout.

Change-Id: Ib3029ea187ba18832f5f8722d5284692347682bf