BUG#37048966: When a generated query fails, mysqltest should print the interpolated query string
* Background

Tests often construct query strings dynamically, store them in
mysqltest $variables, and execute them using eval, `backtick syntax`,
or query_get_value.

Like:

--let $column = 1
--let $query = SELECT value FROM table WHERE column = $column
--let $result = `$query`

When such a query fails, mtr prints the un-interpolated text, like:

mysqltest: At line 193: Query '--let $result = `$query`' failed.

* Problem

Often, the root cause is usually that the query string is not what we
thought it was, so it would be useful to know the interpolated query
that was sent to the server.

* Fix

Make the error message contain the interpolated query.

This patch works for all tested cases, except send_eval/reap, because
the query is sent by `send` but the error is generated by `reap`, at
which time the interpolated query has been forgotten.

Change-Id: Iefbfe419fefc5dfc6155d56ef56d96706e3f6f0e