Bug#37018579: Hypergraph: Ignore first_row_cost when the result is read to the end
The hypergraph optimizer currently preserves candidate subplans with a
high estimated cost if the estimate for the cost of reading a single
row is lower than in the alternative plans. This is done in the hope
that for example a LIMIT operation at a later execution stage does not
have to read the full result and can take advantage of the low cost of
retrieving the first result. It can also be taken advantage of by a
semijoin or antijoin.

However, if there is no LIMIT and no semijoin or antijoin in the
query, the total cost of the query does not benefit from getting the
first row faster from a subplan, as it has to read all rows anyway.

Preserving these candidate makes query planning more time-consuming,
so they should only be preserved if they can actually help finding a
better plan for the query as a whole.

This patch makes CompareAccessPaths() ignore the first row cost for
all subplans that are not under a LIMIT or on the inner side of a
semijoin or an antijoin. It makes the planning of some complex queries
noticeably faster, while still finding a plan with the same estimated
total cost as previously (or, in some cases, a plan with a slightly
lower cost).

The patch also makes CompareAccessPaths() ignore
AccessPath::safe_for_rowid if the query doesn't need row IDs. This
cost dimension was already partly ignored when row IDs were not
needed, by setting safe_for_rowid to SAFE in MakeMaterializePath() if
row IDs were not needed. Moving this logic into CompareAccessPaths()
makes it more effective, as it also applies to materialization paths
created by other functions.

Change-Id: I62da43ffae697e694e2a8dfccc76f5986e23ae21