Bug#37053708: Remove logic that wakes up waiting thread every second to track replication metrics
* Background

We track the time that replication applier threads perform various
wait operations. For most wait operations, the logic to track the time
is:

- When the wait starts, we decrease the metric by the current time,
  which makes it negative.

- When the wait ends, we increase it by the current time. Therefore,
  in conjunction with the previous decrease, this ensures that the
  duration of the wait has been added to the metric.

- When reading the metric, if it is negative, we add the current time
  to the result. This ensures that an ongoing wait is reported
  correctly and with the full precision of the clock.

This logic is captured in the class Time_based_metric.

However, when we track how long time it waits for commit order, we use
a different strategy.  Instead of the above, we use a special wait
function.  The special wait function waits in periods of up to 1
second.  After each period, it increments the metric by the duration
of the period.  The effect is that the waiting time increments in
steps while there is an ongoing wait, where the steps occur
approximately once per second. When the wait completes, the metric
will hold the waiting time with the full precision of the clock.

* Problem

The special wait function was introduced in an earlier development
version where Time_based_metric was inefficient due to using a mutex.
Now it does not use a mutex, so Time_based_metric can be used without
harming performance. Therefore, the special wait function is
unnecessary extra code that we could get rid of if we use
Time_based_metric instead.

Also, if the metric is queried more often than once per second during
an ongoing wait, it will only be observed to increase once per second,
even when the clock precision is higher than that.

* Fix

Use Time_based_metric instead of the special wait function.

Also, use scope guards when measuring time of an operation.

Testing:

- The change logic can only be observed if the metric is queried more
  often than once per second. This would be hard to test with a
  deterministic test case. Therefore, we don't test this.

- But we add a test for another property, that was previously not
  tested: when two workers are waiting concurrently, both increment
  the metric, so the metric increments by a pace twice the wall clock.

Change-Id: I77cd967de33822a048fa4698d1e4ff668411abee