Bug#37031285 regression: crash in get_full_func_mm_tree()
The index on the field we group by in the subquery is required for the
server exit. The outer aggregate has been replaced by a field which to
the optimizer looks like it comes from the subquery's from table and
its m_table_ref which is accessed during optimization (due to the
index) in get_full_func_mm_tree. Since the field refers to a windowing
output tmp table, it m_table_ref is empty (always empty for tmp
tables), so a crash ensues.

The problem here is that the outer aggregate has been replaced by the
Item_field of the windowing output table, introduced in
commit 8bfd77b06:

Bug#35855294 Inconsistent results of SELECT statement with EXISTS clause -
             part II: IN/ALL/ANY [noclose]

for other uses case, described therein. While constructing the
window's frame buffer tmp table CreateFramebufferTable calls
ReplaceMaterializedItems, which sets
LEX::m_splitting_window_expression if we are constructing the frame
buffer. WalkAndReplace checks this value to determine if it should
walk and replace inside a subquery, which is necessary when a subquery
references the *result* of a window function in the same expression,
"splitting".

But in our case, the subquery isn't just part of an expression
*containing* a window function, the subquery is actually an *argument*
of the window function, so in this case we should not do the walk,
since the outer reference is available on the window's input tmp
table, so no rewriting is required. This means the the optimizer will
no longer see the replacement field, but the Item_aggregate_ref, so we
don't get the bug.

Solution: temporarily disable the flag
LEX::m_splitting_window_expression if we dive into a window function's
arguments.

While this solves the repro case, it is possible that similar issues
could pertain to the original uses cases of Bug#35855294 iff there
were indexes present.  This should be investigated separately.

Change-Id: I02c1761c5afe650f32afc6f5986005a96ae265ef