Bug#36322812 - RAPID global variables accept 'OF' as valid bool value Bug#36924845 - RAPID session variables accept 'OF' as valid bool value
The global and session plugin variables are accepting the incorrect
string "OF" as a legitimate bool value. Because the partial matching
argument in find_type() is true, the incorrect string 'OF' partially
matches with bool_typelib's 'OFF'.

The issue does not arise with 'O' (single O), since it is partially
matched with two distinct bool_values {"OFF", "ON"}; however,
find_type() has a check to reject more than one partial matches.

Change-Id: Ic0bde91e0d0f9dbeca39a9459d853ade0bbbda2c