Bug#37080195 Build break with old cmake versions on modern Linux hosts
The build of 8.0 is broken on e.g. Fedora 40 if using cmake-3.11
The problem is that PKG_CHECK_MODULES, which is defined in
Modules/FindPkgConfig.cmake fails to set pkgcfg_lib_TIRPC_tirpc

This is not a problem on el6, where we actually *have* to use cmake-3.11
since it has no TIRPC package. It is a problem only when
explicitly using an old cmake version on a modern Linux version.

Being able to use an old cmake version is important, to ensure we do
not introduce new, unimplemented, cmake features in our codebase.

The fix: use pkgcfg_lib_TIRPC_tirpc if found, otherwise use -ltirpc.

Change-Id: Ibb92d451d8caed0ff87004cc1dde364577187639