BUG#37083501: replication_applier_metrics service not tolerant to missing fields
* Background

The replication_applier_metrics service API is meant to be extensible
while keeping ABI compatibility. In order to add or remove fields in
the future without changing the ABI, the metric values are exchanged
between shared objects using an array where each element is a
<typecode, value> pair. If we need to add a field in the future, we
add a new typecode. If we need to remove a field in the future, we
just don't include an entry with the typecode in the array.

To represent fields that can be either NULL or have a value, the value
is stored in one array entry, and the NULL-ness in another (boolean)
array entry.

* Problem

The component logic to construct the table was not handling missing
fields properly. This is not a problem now, but unless fixed, could
potentially make it impossible to remove fields in the future.

* Fix

Make the component logic detect when a field is missing, and make it
display NULL if the field is nullable, otherwise 0/'' depending on the
data type.

Change-Id: I8a16f4b8dceaac55360165cf5e7abecaf3a02099