Bug #11765514: --COMPACT REMOVES --TZ-UTC IN MYSQLDUMP
--tz-utc, when on (as is the default) does 2 things:
  a) Executes "/*!40103 SET TIME_ZONE='+00:00' */" before reading data
  b) prints "/*!40103 SET TIME_ZONE='+00:00' */;" into the dump file

When --compact is specified, it turns b) off but doesn't turn off a)
This is wrong, as the data are still extracted as if UTC is used, but
actually this is not recorded in the dump file.

Looks like --compact is set to remove all fluff, including the other
version comments too. So, to keep it compatible, this is preserved
and extended to --tz-utc as well.

Now --compact turns off --tz-utc completely and reverses all of its
effects, a) included.

Change-Id: I68584b7fb55b39d71bad342e03732de8ee8884e6