Bug#37079336: Assert failure in Item_eq_base::append_join_key_for_hash_join() with hypergraph
Queries with derived correlated tables could in some cases return
wrong results or exit with assertion failures when using the
hypergraph optimizer.

The problem was that such queries could try to materialize the result
of joining a subset of the tables in the query. The logic in
FinalizePlanForQueryBlock() is not currently able to set up the
materialization correctly before all the tables have been joined
together, and this caused wrong results. Additionally, the hash join
code is not currently prepared for reading input from a temporary
table that does not have its own bit in the table_map data structure,
which caused assert failures in some queries.

The patch fixes these problems by not adding materialization steps in
the query plan before all the tables have been joined together. This
only happened in one case, which was when a sort path was added on top
of a nested loop join with a correlated derived table on the inner
side, and this sort needed to store row IDs, and the result of the
sort would be joined with some other table later. Then a
materialization step would be needed before the sort to provide row
IDs, since the row IDs that point into the correlated derived table
are not valid after the nested loop join.

Since the patch skips the materialization step in this case, it also
has to skip the sorting. Which means that the sorting must instead
happen after all the tables have been joined together, at which point
adding a materialization step is supported by the current code.

Since materialization before all tables are joined is no longer
proposed, the workaround added in bug#36855521 is no longer needed,
and it is removed.

Change-Id: I08be3849b10a730640d94bbdd9bec2e5798b8a4b