Bug#36796094  Column rename in multi-column foreign key returns error
Background:
  WL#14785 introduced support for RENAME COLUMN using ALGORITHM=INSTANT
in 8.0.28. The worklog enforces that ALGORITHM=INSTANT cannot be done if
the column is being referenced by another table due to foreign keys.

Issue:
  The check `ok_to_rename_column` skips checking the last field of the
foreign key constraint when deciding if INSTANT algorithm can be used.

Fix:
  Ensure that all fields of the foreign key constraints are checked for
match when checking if ALGORITHM=INSTANT can be used to perform the
rename

Change-Id: I3105b8eb9242becc9622a0614d50b8f6ae3a611b