Bug#36041256 Router does not check if server version is higher or not
So far the Router never checked the Cluster/ReplicaSet/ClusterSet nodes version.
The (wrong) assumption was that the backward compatibility will never be broken.
This is no longer the case in 8.3 server, which has some feature removals, notably
support for @@group_replication_primary_member system variable that is used
by the pre-8.3 Routers.

Moving forward the Router needs a check that dicards the metadata from the
servers that are highier version than the Router.

This patch adds this check. The Router fails to bootstrap if the version of
the server used for bootstrap is highier than Router version (only major and minor
version is checked, the highier patch version is still supported).
Also during the regular metadata refresh Router checks the server version and
reject the metadata from usupported server version.

Change-Id: Ie2626bd04e86358f4eecb6f6dcc96a8696190fdf