Bug#37057380: Wrong results with sorting or hash join reading from single-record index lookups [1/4, sort, noclose]
Part 1, sorting.

Wrong results were seen in some queries which sorted the result of a
join which accessed one of the tables with a single-row index lookup.
The wrong results were seen if the sort was performed multiple times
in the query (typically because the sort was in a correlated
subquery).

This part of the bug was only seen with the hypergraph optimizer. The
old optimizer tends to add a materialization before sorting the
results of a join, and the materialization avoided this part of the
bug.

The problem is that EQRefIterator, which is used for single-row index
lookup, has a cache that avoids traversing the B-tree again if the
value being looked up has not changed. This requires the
table->record[0] buffer to be unchanged from the previous call to
EQRefIterator::Read(). If an iterator above has changed the contents
of the buffer, the cache is corrupted and the next call to
EQRefIterator::Read() might return wrong values.

The bug is fixed by making SortingIterator::Init() invalidate the
cache in all single-row index lookups below it, so that the first read
from them will be forced to read from the index instead of returning
the cached value.

Change-Id: I7952d4f298aa8face4cbe0ad35ce4029750e180b