Bug#36004394  Skip instant DDL for empty tables
Description:
  ALTER TABLE .. ADD/DROP COLUMN bumps the row version when using
ALGORITHM=INSTANT but performs the DDL instantly. In comparison,
ALGORITHM=INPLACE spends more time as it does a table rebuild. The
default algorithm for ALTER TABLE .. ADD/DROP COLUMN is INSTANT

Enhancement Request:
  ALTER TABLE .. ADD/DROP COLUMN should use INPLACE as the default
algorithm if the table is empty.

  When ADD/DROP COLUMN uses INSTANT, it bumps the row version by 1. As
the version stored in the row is just one byte, there is a limit on the
maximum row version supported. Once this limit is hit, INSTANT is no
longer possible and table must be rebuilt which will reset the limit. As
the rebuild updates every row, there is a larger cost of rebuilding a
large table. Thus, we would prefer to rebuild the table when it is
small or empty to save the row versions for later when it is really
needed. Furthermore, when the table has no records, INPLACE is just as
fast as INSTANT.

  The scenario of ADD/DROP COLUMN on empty table often arises in
automatic deployments - they start with an empty table and then apply
incremental changes to its schema to bring it to the latest shape used
by the application. Thus, using INPLACE over INSTANT will save the row
versions.

Fix:
  This commit modifies the behavior to perform the DDL using INPLACE
algorithm by default when the table has no records. This is checked by
looking at the root page of the first index (primary key or clustered
index) and checking if it is indeed empty. This change only applies to
the default algorithm: The DDL is performed instantly if
ALGORITHM=INSTANT is explicitly provided

Note:
  For small tables in which the records fit within few pages, the speed
gained by using INSTANT over INPLACE is negligible. Thus 3 scenarios
arise where INPLACE can be preferred over INSTANT:
  [1] Primary index has no pages
  [2] Primary index has just one page
  [3] Primary index has "few" pages
This commit chooses [1] to keep things simple. Therefore, DDL uses
ALGORITHM=INSTANT by default if the table contains delete marked records

Change-Id: Id0b1ecef3e0fb0a1e5d2864751a8411e5b7fe592