Bug#36997712: Hypergraph : Result set mismatch               (1 row vs empty set)
When finding replacements while propagating equalities
for a where condition, if a join condition is placed
as part the WHERE clause (could happen with semijoin
transformations), care should be taken to find the
replacements from the same side of the join
if all fields in the condition come from one side
of a join. This is to ensure that such conditions
could be pushed down later. For the case in the
bugpage, replacement for the field in the
semijoin condition is found from a table that
is on the other side of the join that the semijoin
is part of. As a result, the semijoin condition
was not attached to the semijoin resulting
in wrong results. Solution is to use the same
logic that was used earlier for join conditions
i.e to find replacements from the same side of
a join.

Change-Id: I6502c964d7935c5fc221a0db3ae354c6249c9cf9