Bug#36991399: mysqld with signal 6 at option_tracker::deinit()
INSTALL and UNINSTALL COMPONENT aim at ending the transaction
automatically.

However, when they do this they do not release the transaction and
statement MDL locks that might have been taken during the execution
of the component init and deinit methods.
They also acquire (before calling init/deinit) and release
(when the transaction is over) a component service that does the actual
load and unload.

This can lead to the following deadlock with INSTALL and UNINSTALL
running in a loop over the option tracker component:

1. INSTALL COMPONENT has ended its transaction and has released the
dyloader lock (but hasn't yet released the MDL locks over
the performace schema table it has just registered) and
is trying to take a registry lock to release the persisted dynamic
loader service handle it has acquired when it started.
2. At this exact time, in a different thread, UNINSTALL COMPONENT, while
holding the dynamic loader and the registry locks, calls the
option_tracker deinit method.
This method tries to deregister a performance schema table. This
requires that an exclusive transaction level metadata lock over the
table is taken. But, since the INSTALL COMPONENT thread hasn't released
its MDL lock yet, a deadlock occurs.

Fixed by making sure the transaction and statement level MDL locks
are released immediately after the INSTALL/UNINSTALL COMPONENT
transactions are committed or rolled back but *before* the service
handle acquired by INSTALL/UNINSTALL is released.

A test case, while possible, would require a lot of hand-crafted
DBUG_EXECUTE_IF code, and is hence avoided as a bit too unstable.

Change-Id: I109a6cd3a70b62e57c80318c61ce21732e67f48a