Bug#37074489 stored procedure is producing different results in different attempts
The error occurs also in the recursive CTE query if used in a prepared statement;
in the first execution the result is correct, but wrong in subsequent executions.

Tracing revealed that the creation of the materialized tmp table
during optimization was made using the wrong key, so execution found
no row matching its where clause againt the constant in this part of
the query:

   SELECT type,
              COUNT(*) AS count
   FROM t
   WHERE batch_uid = 549285

where t is the recursive cte.

The second execution uses index on "type" instead of on "batch_uid"
for materialized table "t" (and also wrong index for materialized
"tran": "refguid' instead of "batch_uid".

This happens because TABLE::s->key_parts is not reset when the tmp
table is cleaned up in 'cleanup_tmp_tables'. s->keys is reset to 0,
but s->key_parts is not touched. This upsets the logic in
TABLE::add_tmp_key, so we end up with the wrong index (the value for
s->key_parts ends up one too high in the second execution).

Solution: reset TABLE_SHARE::key_parts as well in 'cleanup_tmp_tables'.

Change-Id: I556f041f4476817208a6b8beb179ddaa44fd712f