Bug#36678321: Hypergraph: Sort-ahead doesn't provide row IDs to later sort that needs them [2/2, cleanup]
Additional cleanup for handling of row IDs in UPDATE and DELETE
statements in the hypergraph optimizer.

The hypergraph optimizer enables storing of row IDs in hash join or
sort access paths that access a target table of the UPDATE or DELETE
statement. It does this during planning, before the final plan has
been chosen, so it could spend time unnecessarily on subplans that are
not used in the end. Also, enabling row IDs in a plan also enables it
in its subplans, which may be shared with another candidate that does
not need row IDs, so the final plan may have row IDs enabled in
subplans that don't actually need it.

Fixed by moving the enabling of row IDs to
FinalizePlanForQueryBlock(), so that it's only performed on the final
plan.

Change-Id: I2982f7f62ba18dc60fca1a5e8246cb0f0b3b454b