Bug#37129808: Disable some too strict clang-tidy checks
Disable these two clang-tidy checks:

  misc-no-recursion
  misc-use-anonymous-namespace

The style guide does not discourage recursion, and it does not state
any preference to anonymous namespaces over use of the static keyword.
Since there are lots of recursive functions, and the static keyword is
used a lot, these checks generate too much noise in the clang-tidy
reports.

Additionally, the cppcoreguidelines-avoid-do-while check is configured
so that it ignores macros, as using do-while in macros is a common
idiom, and it is not useful to warn each time such a macro is invoked.

Change-Id: I82ac7f1d945a08435db9278630769091b7f288a9