Bug#37057380: Wrong results with sorting or hash join reading from single-record index lookups [2/4, hash join, noclose]
Part 2, hash join.

This part of the bug affects both the hypergraph optimizer and the old
optimizer.

Wrong results were seen in some cases when the input tables to a hash
join included a table accessed with a single-row index lookup. It is
caused by the hash join changing the contents of table->record[0],
which corrupts the cache in EQRefIterator.

Fixed by invalidating the cache in EQRefIterators below the
HashJoinIterator each time HashJoinIterator is re-initialized.

Change-Id: I01c00191b10f129ad21f282cf674b4eafdf06a62