Bug#36846567 Inplace ALTER TABLE might cause lost rows if concurrent purge
Bug#36261480 alter table tablename engine=innodb cause data loss
Bug#35303494 Data will be lost when the table structure is rebuilt.

Problem:
--------
ALTER TABLE which rebuilds InnoDB table using INPLACE algorithm might
sometime lead to row loss if concurrent purge happens on the table
being ALTERed.

Analysis:
---------
Innodb uses the parallel read infrastructure to scan the index(es) in
parallel through worker threads for ALTER TABLE INPLACE stmt since
MySQL 8.0.27 [WL#14283].

The worker threads scan the records in the subtrees of index(es).
The scan needs to be paused from time to time to give a way to
concurrent actions on the index(es). B-tree index(es) is scanned by
bulk inserter implementation, it  has granularity of the scan pause
at the record level. Spatial index(es) is scanned by the batch inserter
implementation that has granularity of the scan pause at the page level.
Both of the implementations, save the record position in a cursor,
commit the Innodb mini-transaction that releases the page latches it
holds and, restore the record position.

The problem reported in the bug happen at the Bulk inserter while
scanning the B-tree indexes. It works as following : InnoDB maintains
an in-memory key buffer to process the records. Records are added to
the key buffer one by one. The scan is paused when this key buffer is
full. Innodb first saves the position of the previous record
(which was added to the key buffer successfully) to the cursor, releases
the page latches, inserts the key buffer into clustered index, clears
the key buffer, copies the current record into the key buffer and,
restores the cursor. Refer Builder::bulk_add_row() for more details.

Notice that when key buffer is full, the cursor points to the valid
user record position but PCursor::savepoint() wrapper saves the previous
record position and PCursor::restore_position() wrapper restores the
cursor to current record even though underlying persistent cursor APIs
are capable to save/restore the record position correctly. It turns out
this was done to handle the situation where cursor reached to the end of
user records on a page in case of, batch inserter and in the
traverse_recs() method.

Above analysis indicates that current PCursor wrapper APIs are trying
to handle multiple scenarios that make it difficult to implement them
correctly. Instead, we must separate the APIs such that they do
specific task and also neatly.

The idea is to define the savepoint() and restore() APIs for bulk
inserter in this patch and improve the APIs for batch inserted and
traverse_recs() through a separate follow up patch.

Fix:
----
Introduced following two wrapper APIs:

1. savepoint_user_record() :
  Saves the position of a valid user record. This API assumes
  that caller must not be position the cursor at infimum or supremum.

2. restore_user_record_lower_or_equal() :
  Restores the cursor to the saved record position. If the original
  record was removed then restore the cursor to the next largest record
  position which was smaller or equal to the current record. If all
  records were removed then restore the cursor to the infimum; this must
  be fine if the next step of the caller is to advance the cursor.

These APIs rely on underlying persistent cursor(btr_pcur_t)
implementation that handles the cases that bulk inserter cares about.

The test case in this bug fix is based on contribution from Mr Dmitry
Lenev at Percona. We thank you for the contribution to MySQL.

We also thanks Mr LeYuan Zhong and Mr yuanyue Zheng for their
contributions in other reported bugs.

Change-Id: If4b4856940f252da5b6357d016b54323b09bd289