Bug#36808088 Inplace ALTER TABLE might fail with duplicate key error
Problem:
--------
ALTER TABLE which rebuilds InnoDB table using INPLACE algorithm might
sometimes fail with duplicate key error due to a non-duplicate record
inserted concurrently while rebuild was paused that is, page latch was
released.

Analysis:
---------
InnoDB maintains an in-memory key buffer to process the records. The
iteration is paused when this key buffer is full. It first saves the
position of the previous record (which is added to the key buffer
successfully) to the cursor through PCursor::savepoint(), releases the
page latches, inserts the key buffer into clustered index, clears the
key buffer, copies the current record into the key buffer and, restores
the cursor through PCursor::restore_from_savepoint().

PCursor::restore_from_savepoint() uses another helper restore_position()
that first restores the record position to previous record and
forwards it to next record since m_rel_pos remains to BTR_PCUR_ON.
In the meantime, if another records is inserted to the page right after
the previous record position is saved, then this restore logic moves the
cursor to the newly added record that is positioned before the current
record that is copied to the key buffer.
That means when the control reaches back to traverse_recs(), it
moves forward the cursor position to a record which is the same record
that is already added to the key buffer. This leads the duplicate key
error reported in this bug.

Now, we already established in the previous commit (Bug#36846567 fix)
that current PCursor::savepoint() and PCursor::restore_from_savepoint()
APIs try to handle multiple scenarios that breaks the assumptions of
Bulk inserter. That commit use a new pair of APIs for save and restore
during bulk inserter. The new APIs do not suffer from the problem
described above. Therefore, Bug#36846567 patch, fixes this issue as
well.

Fix:
----
Problem is already fixed through Bug#36846567 fix. This commit adds the
test scenario which is based on the contribution by Mr Dmitry Lenev at
Percona. We thank you for contribution to MySQL.

Change-Id: Iba7ddf0f5b22f36526d06a56654630621d9390d1