Bug#36155195 Autotest: testNodeRestart -n TransStallTimeout T1 fails
Backport to 7.6

Problem

Testcase was added as part of fix for
  Bug#22602898 NDB : CURIOUS STATE OF TC COMMIT_SENT / COMPLETE_SENT TIMEOUT HANDLING

That fix modified TC timeout handling to avoid the 'switch to
serial' behaviour when detecting a timeout, as it can result
in excessive slowdown for large transaction commits, leading
to long + large epochs, replication problems etc.

Test testNodeRestart -n TransStallTimeout tested this fix by stalling
a transaction in the commit or complete phase and checking that it
was *not* unblocked (and committed) by TC timeout switching to
the serial commit protocol.

However there was an issue with transient transaction states
which were not directly handled as part of node failure, and
so relied on the non node-failure timeout handling to cleanup.

This was detected and fixed in :
  Bug#36028828 testNodeRestart -n MixedReadUpdateScan failures

That fix uses the difference between the TC's global 'failure number'
counter and an ApiConnectRecord's local failure number to detect
when a node failure has occurred since a transaction started, to
detect and cleanup transactions which were in transient states when
the node failure was detected.

However it is observed that testNodeRestart -n TransStallTimeout
fails occasionally in the COMPLETE phase - a transaction blocked
in the COMPLETE phase is unblocked by normal COMPLETE timeout
handling.

Investigation shows that the problem is that in some cases the
ApiConnectRecord's failure number is not initialised.

This is because the COMPLETE phase uses a different ApiConnectRecord
(Copy record) than the PREPARE + COMMIT phases, to allow the PREPARE
+ COMMIT record to be used for a new transaction concurrently with
COMPLETE.

Solution

Ensure that the Copy ApiConnectRecord has an initialised Failure
number.

Change-Id: I725d80bb3230eeeeb3c3e357943738dfdfc3f434