Bug#37142400: Move test case with binary content to separate test file
Tests main.sp-external and mle.mle_parser_js tests contain a test case
with binary content.  This causes gerrit to not show changes to the
test and result files.  We should move this test case to a separate
test file.

Also, since the test case does not depend on the MLE component, we can
remove it from the mle test suite.

We also use this opportunity to rename "mle_" prefix from the mle
parser test names since, given that it repeats name of the test suite,
it is redundant.

Change-Id: I416493dbfeb417694ba31b4355ed0183503be3e6