WL#16235 - offload secondary-engine queries that insert into tables with triggers - MySQL server part
This patch adds a new member to Query_tables_list, has_stored_functions.
This variable is used instead of uses_stored_routines() check in
all the places that check offloads for the secondary engine.

To streamline updating this variable, an additional stored-routine
type argument is added to sp_add_used_routine.

As a consequence, we now don't block offload of select queries from
insert into select statements to secondary engine even when the
destination table has a defined trigger.

Change-Id: I59662ab8e87a893a50de8d25b76108322a244b6e