Bug#37139618  perfschema.idx_compare_mems_by_thread_by_event_name test failing rarely
Symptom:
Debug build o mysqld occasionally fails on assertio in
ddl::Key_sort_buffer::serialize (ddl0buffer.cc):
ut_ad(ptr > io_buffer.first);

Cause:
The code around this assertio supports the case where data left in the
buffer after serializing the records would overrun the buffer if aligned
to IO_BLOCK_SIZE. In this case all complete blocks of data are persisted
to make room in the buffer for the final block. The scenario occurs
when the length of that left over data is a multiple of IO_BLOCK_SIZE;
in this case all of it is persisted, and nothing is left. The assertion
does not anticipate that situation, although the code supports it
correctly.

Fix:
Assertion is changed to allow for this scenario - after persisting
the data the buffer may be empty, and "past the end" pointer will
be equal to the start of the buffer.

Change-Id: Icbaef197123f0fc85dbdba7c0259e3dcc9352920