Bug#37171855 MTR: Add support to configure secondary engine plugin persistence
The patch adds a hook in MTR driver to prepare and set configuration
related to the secondary engine plugin as a pre-requisite to setup
persistence.

Also, an env variable $SECONDARY_ENGINE_PLUGIN_PERSISTENCE_ENABLED
has been introduced for tests to check if persistence has been
configured and a mechanism has been introduced during mysqld restart
within a test which allows it to skip waiting for secondary engine’s
status to be active. It can be used by setting the below variable -
$SKIP_SECONDARY_ENGINE_BOOTSTRAP_WAIT_ON_MYSQLD_RESTART

Change-Id: I7a181970cfd5e19ecdd97add9c00121008cc8e9b