Bug#37177744: Wrong ordering of outer join results with the hypergraph optimizer [2/2]
The hypergraph optimizer sometimes incorrectly elided sorting in
queries with outer joins, and returned results that were not ordered
according to the ORDER BY clause.

This happened because it collected functional dependencies from
predicates evaluated on the inner side of an outer join, without
taking into account that rows could be NULL-complemented by the outer
join, so that the functional dependency might not hold after the outer
join has been performed.

For example:

  SELECT t2.a FROM t1 LEFT JOIN t2 ON t1.a = t2.a AND t2.a = 2 ORDER BY t2.a

The hypergraph optimizer elided the ORDER BY because t2.a was assumed
to be constant. This assumption did not hold because the outer join
could return rows with t2.a = NULL in addition to rows with t2.a = 2.

Fixed by not collecting functional dependencies from item = const
predicates on the inner side of outer joins. Such predicates do not
guarantee that "item" always has the same value. Functional
dependencies from item = item predicates are still collected, as they
would hold even after NULL-complementing the row (both items will be
NULL).

The mtr test cases show that the queries that previously returned rows
in the wrong order, now returns correctly ordered results.

The unit test cases show that the fix is not too strict, and doesn't
prevent collection of all functional dependencies on the inner side of
outer joins.

Change-Id: Iec8083defdc83d3742a3a2a72fc5de4f200a6760