Bug#37177744: Wrong ordering of outer join results with the hypergraph optimizer [1/2, noclose]
Initial refactoring to make the fix for the bug simpler.

The JoinHypergraph struct has two bitmaps which tell which tables are
on the inner side of some outer join, antijoin or semijoin. The
information is conflated, as there is one
table_map (tables_inner_to_outer_or_anti) giving the union of tables
inner to some outer join or antijoin, and one
NodeMap (nodes_inner_to_semi_or_anti) giving the union of tables inner
to some semi join or anti join.

Change it to have three NodeMap members instead, and have information
about only one join type per bitmap: nodes_inner_to_outer_join,
nodes_inner_to_semijoin and nodes_inner_to_antijoin.

With this change, one can now easily check if a table is on the inner
side of for example an outer join. The existing code that checks if a
table is on the inner side of one out of two join types, can still do
that by checking against the union of two of the bitmaps.

Change-Id: I1f82884f56513c8051d85425cd6a61dbb2d11955